new layout for filesystem
