+++ others/simulation_skeleton.py
@@ -1,32 +1,29 @@
+import os
+import sys
+PROJECT_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir))
+sys.path.append(PROJECT_ROOT)
+
 import simpy
 import pygame
 from abc import ABC, abstractproperty
-
-
+from train_simulation import Moving, Entity
 
 class Train(Moving):
     image = 'Gene youre awesome'
-    position_x = 0
-    position_y = 0
 
 
 class Station(Moving):
     image = 'Gene is a genius'
-    position_x = 0
-    position_y = 0
 
 class Person(Entity):
     image = 'Gene is so hot'
-    position_x = 0
 
 class Railway(Entity):
     image = 'Gene is '
-    position_x = 0
 
 
 def main():
     pygame.init()
-
     # Set up the drawing window
     screen = pygame.display.set_mode([1000, 700])
     return
