+++ train_simulation/Central_comunication.py
