+++ train_simulation/Entity.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractproperty
+
+class Entity(ABC):
+    @abstractproperty
+    def image(self):
+        pass
+
+    def get_figure(self):
+        return self.image
