+++ train_simulation/Person.py
@@ -1,0 +1,3 @@
+
+class Person:
+    pass
